Jewel loan and term deposit rdlc report added

+++ Infin8.Coapp.Dto/rptJewelLoanLedgerOrnmentsSub.cs
@@ -2,7 +2,7 @@
 {
     public  class rptJewelLoanLedgerOrnmentsSub
     {
-        public int Loan_Id { get; set; }
+        public decimal Loan_Id { get; set; }
         public string? JLO_Name { get; set; }
         public int JLO_Nos { get; set; }
     }
