+++ Infin8.Coapp.Models/SBCA_Schemes.cs
@@ -8,8 +8,8 @@ namespace Infin8.Coapp.Models
         [Key]
         public int Scheme_Id { get; set; }
         public string? SBCA_Name { get; set; }
-        public int SBCA_Led_Id { get; set; }
-        public int SBCA_Int_Led_Id { get; set; }
+        public decimal SBCA_Led_Id { get; set; }
+        public decimal SBCA_Int_Led_Id { get; set; }
         public bool SBCA_Delete { get; set; }
         public int InterestCalcPeriod { get; set; }
         public string? BrCode { get; set; }
